Add local server to jome, and new `jome-wake-up` to show jome's window
This new feature makes jome optionally listen for a "wake up" signal
using a local server (Unix socket on this platform). When you pick an
emoji, jome does not quit, but hides the window. Then you can use the
new `jome-wake-up` command to show the hidden jome window again to pick
another emoji.

The purpose of this is that waking up jome is much more snappy than
starting it from scratch. The `jome-wake-up` command, which does not
use Qt GUI/Widgets, is also very fast to execute.

You can start jome as a server with the `-s` option which specifies the
server name. On Unix, this is going to be a socket file in `/tmp`. If
the file already exists, the server might not work, but jome will run
anyway. It could be more safe to remove this file before starting jome
in server mode:

    rm -f /tmp/mein-jome
    jome -s mein-jome

jome starts but does not show its window. To show it:

    jome-wake-up mein-jome

The jome window shows up immediately. Once you pick an emoji, jome hides
the window again. You can make jome quit gracefully (this removes the
socket file):

    jome-wake-up -q mein-jome

Because jome only hides the window and does not quit in this mode, it's
harder to connect it to a script which only waits for it to quit and
uses whatever tool to type the printed emoji or codepoints. I guess this
could be possible by redirecting its standard output and waiting for a
new line. But I'm also introducing an option to execute a command
directly from jome. The command string you pass is interpreted like a
shell does so you can have arguments in there. jome appends to this
command what it would normally print, without a newline (forced), as
one or more extra arguments. For example, with:

    jome -c 'my-command --hello' -f cp

When you pick an emoji, say "Woman detective", jome executes:

    my-command --hello 1f575 fe0f 200d 2640 fe0f

jome hides its window, waits 20 ms, and executes the command. The idea
is that, if your current keyboard focus is within a text box within
a given window, you run `jome-wake-up` with a key binding to show the
jome window, and when you pick an emoji, because the window disappears,
the window manager puts the keyboard focus back to the previous window,
and then the typing tool command executes. This works great on my side.

To accomodate xdotool, which accepts keys with the `U` prefix, not the
`U+` prefix, I'm removing the `ucp` format (which prepended `U+` to each
codepoint) and introducing the `-p` format with which you can have a
custom prefix for each codepoint. So the old `-f ucp` option is now
`-f cp -p U+`.

So the jome command I use now is:

    jome -s jome.socket -f cp -p U -c 'xdotool key --delay 20'

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>